# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/reviews" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getReviews",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "review-controller" ],
        "operationId" : "createReviews",
        "parameters" : [ {
          "name" : "actorRole",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "DELIVERY", "RESTAURANT", "ADMIN" ]
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateReviewRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListReviewDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/orders/{orderId}/eligibility" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getEligibility",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "actorRole",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "DELIVERY", "RESTAURANT", "ADMIN" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReviewEligibilityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/received" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getReceivedReviews",
        "parameters" : [ {
          "name" : "actorRole",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "DELIVERY", "RESTAURANT", "ADMIN" ]
          }
        }, {
          "name" : "outletId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewReceivedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/me" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getMyReviews",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedModelReviewDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/aggregates" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getAggregates",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          }
        }, {
          "name" : "entityIds",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAggregateBatchDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/aggregate" : {
      "get" : {
        "tags" : [ "review-controller" ],
        "operationId" : "getAggregate",
        "parameters" : [ {
          "name" : "entityType",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          }
        }, {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReviewAggregateDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "CreateReviewRequest" : {
        "required" : [ "entries", "orderId" ],
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entries" : {
            "maxItems" : 20,
            "minItems" : 0,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewEntryRequest"
            }
          }
        }
      },
      "ReviewEntryRequest" : {
        "required" : [ "entityId", "entityType", "rating" ],
        "type" : "object",
        "properties" : {
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "rating" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "ApiResponseListReviewDetailDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewDetailDto"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ReviewReceivedDto" : {
        "required" : [ "authorRole", "createdAt", "entityId", "entityType", "id", "rating" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "authorRole" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "RESTAURANT", "DELIVERY" ]
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedModelReviewReceivedDto" : {
        "required" : [ "content" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewReceivedDto"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
      "ApiResponsePagedModelReviewReceivedDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedModelReviewReceivedDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ReviewDetailDto" : {
        "required" : [ "authorRole", "createdAt", "entityId", "entityType", "id", "orderId", "rating", "userId", "visibility" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string"
          },
          "authorRole" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "RESTAURANT", "DELIVERY" ]
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE" ]
          },
          "authorDisplayName" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponsePagedModelReviewDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedModelReviewDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageMetadata" : {
        "required" : [ "number", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "PagedModelReviewDto" : {
        "required" : [ "content" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewDto"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
      "ReviewDto" : {
        "required" : [ "createdAt", "entityId", "entityType", "id", "rating" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comment" : {
            "type" : "string"
          },
          "authorDisplayName" : {
            "type" : "string"
          },
          "authorRole" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "RESTAURANT", "DELIVERY" ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseReviewEligibilityDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ReviewEligibilityDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ReviewEligibilityDto" : {
        "required" : [ "orderId", "reviewable", "targets" ],
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reviewable" : {
            "type" : "boolean"
          },
          "reason" : {
            "type" : "string",
            "enum" : [ "ORDER_NOT_FOUND", "NOT_YOUR_ORDER", "ORDER_NOT_DELIVERED", "REVIEW_WINDOW_CLOSED", "TARGET_NOT_ON_ORDER", "ALREADY_REVIEWED", "DUPLICATE_ENTRY", "ROLE_TARGET_NOT_ALLOWED", "SELF_REVIEW" ]
          },
          "reasonDetail" : {
            "type" : "string"
          },
          "windowClosesAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "targets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewTargetDto"
            }
          }
        }
      },
      "ReviewTargetDto" : {
        "required" : [ "alreadyReviewed", "displayName", "entityId", "entityType", "visibility" ],
        "type" : "object",
        "properties" : {
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "visibility" : {
            "type" : "string",
            "enum" : [ "PUBLIC", "PRIVATE" ]
          },
          "alreadyReviewed" : {
            "type" : "boolean"
          },
          "existingRating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "existingComment" : {
            "type" : "string"
          },
          "existingReviewedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponsePagedModelReviewDetailDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedModelReviewDetailDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedModelReviewDetailDto" : {
        "required" : [ "content" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewDetailDto"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
      "AggregateBatchDto" : {
        "required" : [ "aggregates", "entityType" ],
        "type" : "object",
        "properties" : {
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "aggregates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReviewAggregateDto"
            }
          }
        }
      },
      "ApiResponseAggregateBatchDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AggregateBatchDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ReviewAggregateDto" : {
        "required" : [ "averageRating", "entityId", "entityType", "totalReviews" ],
        "type" : "object",
        "properties" : {
          "entityType" : {
            "type" : "string",
            "enum" : [ "RESTAURANT", "DRIVER", "PRODUCT", "CUSTOMER" ]
          },
          "entityId" : {
            "type" : "string"
          },
          "totalReviews" : {
            "type" : "integer",
            "format" : "int64"
          },
          "averageRating" : {
            "type" : "number"
          }
        }
      },
      "ApiResponseReviewAggregateDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ReviewAggregateDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    }
  }
}