# src/main/java/com/fooddelivery/reviews/dto/ReviewAggregateDto.java
package com.fooddelivery.reviews.dto;

import com.fooddelivery.common.enums.ReviewEntityType;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;
import java.math.BigDecimal;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ReviewAggregateDto implements Serializable {
    @jakarta.validation.constraints.NotNull
    private ReviewEntityType entityType;
    @jakarta.validation.constraints.NotNull
    private String entityId;
    @io.swagger.v3.oas.annotations.media.Schema(requiredMode = io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED)
    private long totalReviews;
    @jakarta.validation.constraints.NotNull
    private BigDecimal averageRating;
}